finished tutorial
I ended up having to copy paste from the github post linked in the video, there was a part of my code that was giving an error and I wasn't about to go back from the start of the video and re type everything. I've already watched the video in it's entirety and followed along so I think it's justified.